requirements.txt
# PDF Parsing
llama-cloud-services>=0.1.0

# AI/NLP
anthropic>=0.39.0

# Utilities
python-dotenv>=1.0.0

